pyproject.toml
[project]
name = "adventofcode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pathlib>=1.0.1",
    "pillow>=12.0.0",
    "seaborn>=0.13.2",
]
